+++ packages/plugin-e2e/src/models/index.ts
@@ -1,8 +1,8 @@
-export { AnnotationEditPage } from './AnnotationEditPage';
-export { AnnotationPage } from './AnnotationPage';
-export { DashboardPage } from './DashboardPage';
-export { DataSourceConfigPage } from './DataSourceConfigPage';
-export { GrafanaPage } from './GrafanaPage';
-export { PanelEditPage } from './PanelEditPage';
-export { VariableEditPage } from './VariableEditPage';
-export { VariablePage } from './VariablePage';
+export { AnnotationEditPage } from './pages/AnnotationEditPage';
+export { AnnotationPage } from './pages/AnnotationPage';
+export { DashboardPage } from './pages/DashboardPage';
+export { DataSourceConfigPage } from './pages/DataSourceConfigPage';
+export { GrafanaPage } from './pages/GrafanaPage';
+export { PanelEditPage } from './pages/PanelEditPage';
+export { VariableEditPage } from './pages/VariableEditPage';
+export { VariablePage } from './pages/VariablePage';
